changed back to normal

+++ app.py
@@ -37,13 +37,6 @@ def add_user_to_g():
         g.user = User.query.get(session[CURR_USER_KEY])
     else:
         g.user = None
-    
-    if not g.user and request.endpoint not in ("do_login", 
-        "do_logout", "signup", "login", 
-        "logout", "list_users", "users_show", 
-        "messages_show", "homepage", "add_header"):
-        flash("Access unauthorized.", "danger", "homepage")
-        return redirect("/")
 
 
 def do_login(user):
@@ -165,7 +158,9 @@ def users_show(user_id):
 def show_following(user_id):
     """Show list of people this user is following."""
 
-    #####
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     user = User.query.get_or_404(user_id)
     return render_template('users/following.html', user=user)
@@ -175,7 +170,9 @@ def show_following(user_id):
 def users_followers(user_id):
     """Show list of followers of this user."""
 
-    #####
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     user = User.query.get_or_404(user_id)
     return render_template('users/followers.html', user=user)
@@ -185,7 +182,9 @@ def users_followers(user_id):
 def add_follow(follow_id):
     """Add a follow for the currently-logged-in user."""
 
-    #####
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     followed_user = User.query.get_or_404(follow_id)
     g.user.following.append(followed_user)
@@ -198,7 +197,9 @@ def add_follow(follow_id):
 def stop_following(follow_id):
     """Have currently-logged-in-user stop following this user."""
 
-    ######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     followed_user = User.query.get(follow_id)
     g.user.following.remove(followed_user)
@@ -210,7 +211,9 @@ def stop_following(follow_id):
 @app.route('/users/add_like/<int:msg_id>', methods=['POST'])
 def like_message(msg_id):
     
-    #######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     user_id = g.user.id
     user_likes = g.user.likes
@@ -229,7 +232,9 @@ def like_message(msg_id):
 @app.route('/users/<int:user_id>/likes')
 def show_liked_msg(user_id):
     
-    #######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     likes = g.user.likes
     user = g.user
@@ -241,7 +246,9 @@ def profile():
     """Update profile for current user."""
 
     # IMPLEMENT THIS
-    ######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     user = g.user
     form = EditProfileForm(obj=user)
@@ -266,7 +273,9 @@ def profile():
 def delete_user():
     """Delete user."""
 
-    ######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     do_logout()
 
@@ -286,7 +295,9 @@ def messages_add():
     Show form if GET. If valid, update message and redirect to user page.
     """
 
-    #########
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     form = MessageForm()
 
@@ -312,7 +323,9 @@ def messages_show(message_id):
 def messages_destroy(message_id):
     """Delete a message."""
 
-    #######
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     msg = Message.query.get(message_id)
     db.session.delete(msg)
